+++ packages/codegen-ui-golden-files/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.19.1](https://github.com/aws-amplify/amplify-codegen-ui/compare/v2.19.0...v2.19.1) (2023-10-23)
+
+**Note:** Version bump only for package @aws-amplify/codegen-ui-golden-files
+
+# [2.19.0](https://github.com/aws-amplify/amplify-codegen-ui/compare/v2.17.0...v2.19.0) (2023-10-19)
+
+**Note:** Version bump only for package @aws-amplify/codegen-ui-golden-files
+
 # [2.18.0](https://github.com/aws-amplify/amplify-codegen-ui/compare/v2.17.0...v2.18.0) (2023-10-18)
 
 **Note:** Version bump only for package @aws-amplify/codegen-ui-golden-files